# docs/openapi/auth-service.yml
openapi: 3.0.3
info:
  title: Auth Service API 문서
  version: 1.0.0
  description: |
    Auth-server API 문서입니다.
    이 서비스는 로그인, 토큰 갱신, 로그아웃, GitHub SSO, 세션 검증, 내부 auth account 관리를 제공합니다.
servers:
  - url: /
    description: auth-service 직접 호출 기본 경로
  - url: /v1
    description: Gateway v1 prefix가 붙은 경로
...
paths:
...
  /auth/login:
    post:
      tags: [Auth]
      summary: 비밀번호 로그인
      operationId: login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: 로그인이 성공했습니다. access token과 refresh token을 body와 cookie로 반환합니다.
          headers:
            Set-Cookie:
              description: ACCESS_TOKEN과 refresh token cookie입니다.
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /auth/refresh:
    post:
      tags: [Auth]
      summary: access token과 refresh token 갱신
      operationId: refresh
      description: 설정된 refresh token cookie에서 refresh token을 추출합니다.
      responses:
        '200':
          description: 토큰 갱신이 성공했습니다.
          headers:
            Set-Cookie:
              description: rotation된 refresh cookie와 ACCESS_TOKEN cookie입니다.
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    accessTokenCookie:
      type: apiKey
      in: cookie
      name: ACCESS_TOKEN
    ssoSessionCookie:
      type: apiKey
      in: cookie
      name: sso_session
    internalApiKey:
      type: apiKey
      in: header
      name: X-Internal-Secret
...
  responses:
    BadRequest:
      description: 잘못된 요청입니다.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/VoidGlobalResponse'
    Unauthorized:
      description: 인증이 필요하거나 token이 유효하지 않습니다.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/VoidGlobalResponse'
...
  schemas:
...
    LoginRequest:
      type: object
      required: [username, password]
      properties:
        username:
          type: string
          format: email
          example: user@example.com
        password:
          type: string
          minLength: 8
          maxLength: 72
          format: password
          example: Password12!
    TokenResponse:
      type: object
      required: [accessToken, refreshToken]
      properties:
        accessToken:
          type: string
          description: JWT access token입니다.
        refreshToken:
          type: string
          description: JWT refresh token입니다.
...
    VoidGlobalResponse:
      allOf:
        - $ref: '#/components/schemas/GlobalResponse'
        - type: object
          properties:
            data:
              nullable: true
...
    GlobalResponse:
      type: object
      required: [httpStatus, success, message, code]
      properties:
        httpStatus:
          type: integer
          example: 200
        success:
          type: boolean
          example: true
        message:
          type: string
          example: 조회 요청 성공
        code:
          type: integer
          example: 1
        data:
          nullable: true